Remove "-f" option in grant and revoke commands
The "-f" option in grant and revoke commands had intended to provide
a method to bypass the paxos algorithm hence provide the user a totally
manual way to manage the ticket. However, because the usage of "-f"
often confused the user and also crm_ticket commmand in pacemaker can
provide the same functionality, so remove the "-f" option now.

Signed-off-by: Jiaju Zhang <jjzhang@suse.de>