Fixed:correct the calculation of quorum in learner
 Currently, Learner accumulates the vote when booth runs renew. the
number of vote exceeds the total of acceptor after a little. And then,
booth always has the quorum.
 booth use multi-paxos to implement renew. however, booth still dosn't
"instance number". "instance number" can resolve this issue.
 And, when this was fixed, the other problem also happened. Regardless
of learner reach a quorum, booth can't commit a value when a proposer
message is lost. In terms of Paxos, this should commit a value. So, we
move the making of timer to learner(or proposer) from acceptor.

Signed-off-by: Yuichi SEINO <seino.cluster2@gmail.com>