Medium: ticket: increment lease time only when majority held
We need to send a new lease time to others with the broadcast,
but cannot update our local copy until we reached majority.
Otherwise, we'd lose the ticket too late.

The ticket now gets updated with the new RPC which is broadcasted
after the majority was reached. Otherwise, on losing majority
which may have happened in the meantime, some followers may have
different (wrong) information about the ticket expiration time.