High: raft: elections not allowed if the term's still valid (bnc#876300)
If we the term is still valid, don't let others start the
election. It may happen that a follower goes disconnected,
considers the ticket lost after a while and starts election, then
comes back and tries to get the ticket which is still valid and
supported by the majority. The leader, if it loses majority, will
lose the ticket itself and agree to do new elections.