Medium: ticket: reacquire expired granted tickets (bnc#878359)
Ticket expiry is not recognized by pacemaker, hence whether the
granted ticket is old has no effect. Hence, not to surprise the
cluster unnecessarily, try to reacquire even the tickets which
are not valid anymore but granted here.

This may happen if the booth has been down for a long time.