Medium: raft: account for slow servers in elections
On ticket lost, a fast server may start the elections before some
slow server has the ticket expire. The slow server then rejects
the elections claiming that the ticket is still valid. The safest
way to proceed seems to be that the fast server simply backups a
bit and then starts new elections in hope that the slow server in
the meantime expired its ticket as well.