Medium: raft: always accept heartbeats and updates from a valid leader (bnc#882209)
We need to trust the ticket leader.

On rejected elections, it may happen that some members remain
with a (previously rejected) higher term. They need then to
accept the valid lower term which is still at the leader. This
could happen on a say 2-3 split brain or high packet loss
network.