Medium: raft: add the request in reply packets
The server replying to a request now copies that request into the
reply RPC, so that the server that sent the request RPC can match
the ack with the right request.

Another option is to serialize RPC, i.e. add the monotonically
increasing number ID to all RPC, but we don't need that (yet).