Medium: raft: match the request in acks too (bnc#882209)
We send heartbeats and ticket updates in a relatively quick
fashion, one after the other. Since we wait only for the majority
of heartbeats to arrive, a late heartbeat ACK may be interpreted
as an ACK to the ticket update. With this change, the requesting
server can see for which request the Ack has been sent as a reply.