Medium: raft: be more permissive for elections for valid tickets
In case the network is losing packets, it may happen for instance
that a leader wants to step down, but not all servers get the
message. Now, those that didn't would reject elections as they
have a still valid ticket. Here we relax rejecting elections only
in case the requester thinks that the ticket got lost (which is
obviously not true for valid tickets).