test: Allow test running as a root
The main problem was hidden in inability of booth daemon to delete its
lock file because of directory permisions.

Tests creates temp directory as a root but booth daemon drops
privilegies right after creating the lock file.
So booth daemon can ftruncate lock file, but not delete it.

There seems to be two possible solutions:
- Fix test so empty lock file is considered as deleted lock file
- Change permissions of created temp directory

This patch implements later solution.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>