test: Add compatibility with iproute ss command
ensure_boothd_not_running is now exectured only when single instance is
used, but if so, it requires netstat command. Netstat is deprecated for
some time and it is slowly disappearing from distributions, so add
support for iproute ss command.

Parameters for both ss and netstat are same, but sadly output differs so
both netstat and ss output is matched.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>