configure: Make sure without-html_man works
AC_ARG_WITH always calls the action-if-given no matter if --with or
--without is used. Previously we were setting with_html_man to yes in
this action so --without was behaving same way as --with.

Solution is to remove this action and let AC_ARG_WITH do its job and set
with_html_man correctly to yes if --with-html_man is provided, no if
--without-html_man is provided and not set at all if argument is not
used.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>